apps/steering-dashboard/src/requirements.txt
streamlit>=1.50.0
pandas>=2.2.0
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
tabulate>=0.9.0
